bot/requirements.txt
# WhatsApp SLA Bot Dependencies
requests>=2.28.0
redis>=4.5.0
python-dotenv>=1.0.0
Flask>=3.0.0
gunicorn>=21.0.0
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0
